Separate the model from the user_interface file
The model is pretty much exclusively separate from the rest of the code
in ui.ts, this breaks it away.

This also adds tests.
