+++ src/chrome/extension/scripts/context.ts
@@ -1,6 +1,7 @@
 /// <reference path='../../../../../third_party/typings/chrome/chrome.d.ts' />
 /// <reference path='../../../generic_ui/polymer/context.d.ts' />
 
+import ui_model = require('../../../generic_ui/scripts/model');
 import ui_constants = require('../../../interfaces/ui');
 import user_interface = require('../../../generic_ui/scripts/ui');
 import CoreConnector = require('../../../generic_ui/scripts/core_connector');
@@ -12,8 +13,7 @@ export var core :CoreConnector = ui_context.core;
 export var browserConnector :browser_connector.CoreBrowserConnector = ui_context.browserConnector;
 
 export var ui :user_interface.UserInterface = new user_interface.UserInterface(core, ui_context.browserApi);
-
-export var model :user_interface.Model = ui.model;
+export var model :ui_model.Model = ui.model;
 
 ui.browser = 'chrome';
 console.log('Loaded dependencies for Chrome Extension.');
