+++ src/firefox/data/scripts/background.ts
@@ -1,3 +1,4 @@
+import ui_model = require('../../../generic_ui/scripts/model');
 import user_interface = require('../../../generic_ui/scripts/ui');
 import CoreConnector = require('../../../generic_ui/scripts/core_connector');
 import FirefoxCoreConnector = require('./firefox_connector');
@@ -7,7 +8,7 @@ import port = require('./port');
 export var ui   :user_interface.UserInterface;
 export var core :CoreConnector;
 export var browserConnector :FirefoxCoreConnector;
-export var model :user_interface.Model;
+export var model :ui_model.Model;
 var firefoxBrowserApi :FirefoxBrowserApi;
 
 function initUI() {
