+++ src/generic_ui/scripts/ui.spec.ts
@@ -405,67 +405,3 @@ describe('UI.UserInterface', () => {
     });
   });
 });  // UI.UserInterface
-
-describe('user_interface.model', () => {
-  var model :user_interface.Model;
-
-  beforeEach(() => {
-    model = new user_interface.Model();
-  });
-
-  it('Updating global settings correctly updates description', () => {
-    // description is chosen here as just some arbitrary simple value
-    var newDescription = 'Test description';
-
-    model.updateGlobalSettings({
-      description: newDescription
-    });
-
-    expect(model.globalSettings.description).toEqual(newDescription);
-  });
-
-  it('Updating one field of global settings does not change any other fields', () => {
-    var constantString = 'some arbitrary string';
-
-    model.updateGlobalSettings({
-      description: constantString
-    });
-
-    model.updateGlobalSettings({
-      mode: ui_constants.Mode.SHARE
-    });
-
-    expect(model.globalSettings.description).toEqual(constantString);
-  });
-
-  it('Syncing arrays in global settings works fine', () => {
-    var newStunServers = [
-      {
-        urls: ['something.net:5']
-      },
-      {
-        urls: ['else.net:7']
-      }
-    ];
-
-    model.updateGlobalSettings({
-      stunServers: newStunServers
-    });
-
-    expect(model.globalSettings.stunServers.length).toEqual(newStunServers.length);
-
-    for (var i in newStunServers) {
-      expect(newStunServers[i]).toEqual(model.globalSettings.stunServers[i]);
-    }
-  });
-
-  it('Updating global settings does not reassign', () => {
-    var a = model.globalSettings;
-
-    model.updateGlobalSettings({
-      mode: ui_constants.Mode.SHARE
-    });
-
-    expect(model.globalSettings).toEqual(a);
-  });
-});
